+++ Assets/Songs/Song.cs
@@ -5,7 +5,7 @@ public class Song : ScriptableObject
 {
     public AudioClip audio;
     public new string name = "New Song";
-    public int tempo;
-    public int beatsPerBar = 4;
-    public int barsUntilStart = 8;
+    public float tempo;
+    public float beatsPerBar = 4;
+    public float barsUntilStart = 8;
 }
